backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
mysql-connector-python==9.0.0
Pillow==10.0.1
python-magic-bin==0.4.14
Flask-SocketIO==5.3.6
python-socketio==5.9.0
